+++ .gitignore
@@ -1,3 +1,4 @@
 server/server/__pycache__
 client/__pycache__
 voice.mp3
+vision/people
