+++ docker-compose.yml
@@ -1,0 +1,20 @@
+version: '3.4'
+
+services: 
+    brain:
+        build: ./server
+        volumes:
+            - ./server:/usr/src/app
+        ports: 
+            - 5000:80
+        privileged: true
+    computing_vision:
+        build: ./vision
+        volumes:
+            - ./vision:/usr/src/app
+        privileged: true
+    spear:
+        build: ./client
+        volumes:
+            - /.client:/usr/src/app
+        privileged: true
