+++ vision/docker-compose.yml
@@ -1,0 +1,8 @@
+version: '3.4'
+
+services: 
+    server:
+        build: .
+        volumes:
+            - .:/usr/src/app
+        privileged: true
